requirements.txt
attrs==23.2.0
blinker==1.7.0
click==8.1.7
DateTime==5.4
Flask==3.0.2
Flask-Cors==4.0.0
flask-expects-json==1.7.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
rpds-py==0.18.0
setuptools==69.1.1
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
zope.interface==6.2
